Implement MemoryBank, loads, and stores.
This commit involves a few changes to the internal register API.
Our new strategy is to convert the "ISA" registers to our local
interpreter registers as soon as possible, then use visitor patterns on
the `Register` interface to perform register type-dependent behavior.

This lowers the chance we'll forget an instanceof somewhere.
Plus, it's very "Java". Wow.