Add noise to Interpreter's stochastic computing emulation.
The type of noise will have to be thought through more. In a real
accelerator, it won't be "noise" as much as "adherence to the central
limit theorem"; more complex relationship with intended value and the
frame size. Maybe we should put in some Gaussian noise before the actual
computation?

Oh yeah. That would have been _way_ simpler. Ah well.